rename conjob controller folders
