+++ app/core/Engine/controllers/cronjob-controller/constants.ts
@@ -1,0 +1,6 @@
+import { CronjobControllerState } from '@metamask/snaps-controllers';
+
+export const defaultCronjobControllerState: CronjobControllerState = {
+  jobs: {},
+  events: {},
+};
